requirements.txt
anyio==3.6.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.39.3
google-auth==2.17.3
google-colab==1.0.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
influxdb==5.3.1
influxdb-client==1.36.1
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.5
mypy==1.3.0
mypy-extensions==1.0.0
nbformat==5.8.0
nest-asyncio==1.5.6
notion-client==2.0.0
notion-df==0.0.5
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pandas-stubs==2.0.1.230501
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
portpicker==1.5.2
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.9.2
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
reactivex==4.0.4
requests==2.28.2
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
statsmodels==0.13.5
tenacity==8.2.2
tomli==2.0.1
tornado==6.3.1
traitlets==5.9.0
types-pytz==2023.3.0.0
types-requests==2.30.0.0
types-urllib3==1.26.25.13
typing==3.7.4.3
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
zmq==0.0.0
